Skip environment sanity checks in the parallel mode.
Introduce a new environment variable STAP_PARALLEL indicating the
parallel mode. Move the environment sanity checks to a function
environment_sanity_test() and skip it in parallel mode to speed
the testing up.

If we'll want to run this sanity check once, right at the beginning
of the parallel test run, environment_sanity_test() can get wrapped
into a testcase and run right at the beginning of the
installcheck-parallel Makefile target. Skipping this for now.