Use the artifacts dir "concept".
This commit reflects the agreement on using an "artifacts"
directory. The goal is not to have test artifacts (files
produced by tests (e.g. built binaries) stored in weirdly
named directories in /tmp, but to use a clear directory
structure fith self-explanatory dirnames as a storage.

Another goal addressed in this commit is to keep artifacts
on the filesystem after the testsuite finishes. Cleanup
is being done by 'make clean'.

This approach aligns to what the parallelization concept
inspired by the GDB testsuite parallelization already has
for storing logs per testcase.

Nitpick: This commit drops the [info exists env(SYSTEMTAP_DIR)]
bit at several places. Thinking behind this is that this env
var gets set at the testsuite startup in the
setup_systemtap_environment(), so we rely on it.

Changed files:
* Makefile.in: Revert the "tempdir per runtest" part.
* lib/systemtap.exp: Introduce the setup_artifacts_dir().
* systemtap.base/at_var_cu.exp: Testcase update.
* systemtap.base/cache_clean.exp: Ditto.
* systemtap.base/hash_add_buildid.exp: Ditto.
* systemtap.pass1-4/buildok-interactive.exp: Ditto.
* systemtap.server/client_args.exp: Ditto.