Provide TESTS env var for specifying which tests to run in parallel mode.
The new TESTS environment variable exists for the user to pass on the
command line to easily say "Only run these tests." This is similar to
what the GDB testsuite has too and how it implements it.

The "slow" and "serial" tests are still taken into account, so that "slow"
tests are being run first and "serial" tests are being run serially.